The SmxHelper class is much improved and the Smx to Smx2C conversions are working properly.

+++ src/MSetGenP/ApFixedPointFormat.cs
@@ -10,13 +10,13 @@ namespace MSetGenP
 		public ApFixedPointFormat(int limbCount): this(RMapConstants.BITS_BEFORE_BP, limbCount * BITS_PER_LIMB - RMapConstants.BITS_BEFORE_BP)
 		{ }
 
-		public ApFixedPointFormat(int bitsBeforeBinaryPoint, int numberOfFractionalBits)
+		public ApFixedPointFormat(byte bitsBeforeBinaryPoint, int numberOfFractionalBits)
 		{
 			BitsBeforeBinaryPoint = bitsBeforeBinaryPoint;
 			NumberOfFractionalBits = numberOfFractionalBits;
 		}
 
-		public int BitsBeforeBinaryPoint { get; init; }
+		public byte BitsBeforeBinaryPoint { get; init; }
 		public int NumberOfFractionalBits { get; init; }
 
 		public int TotalBits => BitsBeforeBinaryPoint + NumberOfFractionalBits;
